Refactoring UserView
